+++ models/user.go
@@ -7,7 +7,7 @@ import (
 type User struct {
 	gorm.Model
 	Username string
-	Password uint
+	Password uint64
 	Articles []Article
 	Comments []Comment
 }
